project_#5_EIA Generator - Capstone Project/requirements.txt
# =============================================================================
# EIA Generator - Dependencies
# =============================================================================

# Core AI/ML
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langgraph>=0.0.20
openai>=1.10.0

# Search & Retrieval
tavily-python>=0.3.0
chromadb>=0.4.22
sentence-transformers>=2.2.2

# Document Generation
python-docx>=1.1.0
reportlab>=4.0.0
jinja2>=3.1.0
pypdf>=3.17.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Web Framework
streamlit>=1.30.0
fastapi>=0.109.0
uvicorn>=0.25.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0
omegaconf>=2.3.0

# Utilities
rich>=13.7.0
loguru>=0.7.2
tqdm>=4.66.0
httpx>=0.26.0
aiohttp>=3.9.0

# Charts & Visualization
matplotlib>=3.8.0
plotly>=5.18.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Development
black>=23.12.0
isort>=5.13.0
